devfreq: ondemand: simplify the governor
Simple ondemand did some really crazy divisions when
calculation the next frequency.
To me this seemed to be placeholders. Looks like someone
really wanted to make this very basic and simple.
Switching to a smarter decision algorithm to find out the
next frequency, which should be easier to read and maintain.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>